Implement optional rustfmt integration and refactor transformer module
- Added `rustfmt_fmt.rs` for formatting Rust code using rustfmt with a fallback mechanism.
- Introduced `test_support.rs` for shared test fixtures to create minimal Cargo projects.
- Refactored `transformer.rs` by splitting functionality into dedicated modules: `docs.rs`, `expansion.rs`, and `visit_mut_impl.rs`.
- Moved documentation-related functions to `docs.rs` for better organization.
- Implemented external library expansion and module handling in `expansion.rs`.
- Enhanced `visit_mut_impl.rs` to handle AST traversal and transformation.
- Updated tests to ensure formatting and transformation functionalities work as expected.
